[#20] feat: 드롭다운 재사용성 높이기 위해 공통 컴포넌트화
